% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sm_dir.R
\name{check_sm_dir}
\alias{check_sm_dir}
\title{Check sm_dir input for formatting}
\usage{
check_sm_dir(sm_dir)
}
\arguments{
\item{sm_dir}{directory where data files from the SM board or DTAG (e.g., xml and swv files) are stored}
}
\value{
sm_dir formatted as tagtools functions expect (including final / and using / not \)
}
\description{
this function is called by other functions that read SM board data files. It checks the format of the SM path provided by the user. It is not normally called by end users.
}
