% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deglitch.R
\name{deglitch}
\alias{deglitch}
\title{Remove single-sample outliers from sensor data vector}
\usage{
deglitch(x, m = 10, reps = 4)
}
\arguments{
\item{x}{numeric data vector}

\item{m}{Multiplicative factor by which the first difference of the sample data must exceed the mean first difference, to be replaced with preceding value. Default: 10.}

\item{reps}{Number of times to run the detection iteratively (when outliers are replaces, the mean first difference will change...) Default: 4}
}
\value{
x, with single-sample outliers replaced by the preceding value ("outliers" as defined above)
}
\description{
Remove single-sample outliers from sensor data vector
}
