% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_offset_pressure.R
\name{fix_offset_pressure}
\alias{fix_offset_pressure}
\title{Correct a depth or altitude profile for offsets caused by miscalibration and sensor drift.}
\usage{
fix_offset_pressure(p, sampling_rate, interval = 3600, tc = 12 * 3600)
}
\arguments{
\item{p}{A sensor list or vector of depth/altitude in meters}

\item{sampling_rate}{The sampling_rate of p in Hz. This is only needed if p is not a sensor list.}

\item{interval}{is the search interval in seconds that is used to find surfacings or landings. This should be chosen to be a little more than the usual inter-surfacing/inter-landing interval. Default value is 1 hour.}

\item{tc}{is the smoothing time constant in seconds used to filter the surface depth offsets. \code{tc} should normally be several times larger than \code{interval} unless the depth/altitude sensor has a fast drift. Default value is 12 hours. If \code{tc} is less than or equal to \code{interval}, no smoothing is performed.}
}
\value{
A list with 2 elements:
\itemize{
 \item{\strong{p: }} A sensor list or vector of corrected depth/altitude measurements at the same sampling rate as the input data. If the input is a sensor list, the output will also be.
 \item{\strong{offsets: }} is a 2-variable data frame containing a set of \code{times} (column 1) and estimated \code{pressure_offset}s (column 2). Times are in seconds since the first sample in p. Pressure offsets are in metres.
}
}
\description{
This function finds minima in the dive/altitude profile that are consistent with surfacing/landing and smooths these to make a time-varying '0 pressure' offset vector. This tool should be used if there are still pressure offsets after correcting for temperature using manufacturer black-box calibration or e.g. \code{\link{fix_pressure}}.
}
\note{
This function makes a number of assumptions about the depth/altitude data and about the behaviour of animals: First, the depth data should have few incorrect outlier (negative) values that fall well beyond the surface. These can be reduced using \code{\link{median_filter}} before calling \code{fix_offset_pressure}. the pressure offset in the sensor varies slowly and smoothly with respect to the inter-surfacing/inter-landing interval. This function will not be effective at correcting step changes in calibration.
}
