% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_br.R
\name{get_br}
\alias{get_br}
\title{Use magnetometer data to estimate the body rotations from cyclic locomotion movements.}
\usage{
get_br(Ma, sampling_rate, fh, thr = 0.2, axis = "x")
}
\arguments{
\item{Ma}{is the triaxial magnetometer data in the animal frame. Ma can be a sensor structure or a three-column matrix. The sampling rate of Ma must be at least 2x the highest stroke rate of the animal. The magnetometer data can be in any units as long as all three    columns have the same unit.}

\item{sampling_rate}{is the sampling rate of Ma in Hz, and is only needed if Ma is not a sensor structure.}

\item{fh}{is the high-pass filter frequency in Hz to use to separate orientation changes from locomotory strokes. It should be about half of the dominant stroke frequency. Use \code{dsf()} to estimate the dominant stroke frequency.}

\item{thr}{is an optional minimum field strength threshold to prevent errors in the computation. Errors arise if the plane of rotations is nearly perpendicular to the local magnetic field vector. To avoid these, the body rotation signal is replaced with NaN if the field strength in the locomotory plane drops below thr fraction of the total field strength. The default value is 0.2 (i.e., the locomotory plane must have at least 20 percent of the total field strength to compute the body rotations).}

\item{axis}{is an optional indicator that the locomotion is in the x-y plane. The default is \code{axis = 'x'}, that is, the function expects the locomotion to be in the x-z plane (e.g., cetacean swimming) by default. To compute body rotations in the x-y plane (e.g., for pinnipeds and many fish), use \code{ax = 'y'}.}
}
\value{
A one-column matrix containing the body rotation signal in radians. It has the same sampling rate and number of samples as Ma.
}
\description{
For derivation see: Martin López L, Aguilar de Soto N, Miller P, Johnson M
  2016 Tracking the kinematics of caudal-oscillatory swimming: A comparison 
  of two on-animal sensing methods. J Exp Biol 219:2103-2109.
}
\details{
Note: to estimate the stroking rate of small animals for which the specific
acceleration is larger, it may be simpler to find the zero crossings on the 
high-pass-filtered acceleration (z axis for cetaceans, y axis for pinnipeds
 and fish) directly rather than using \code{get_br}.
}
\examples{
ph <- get_br(harbor_seal$M, fh = 0.06, axis = 'y')
# choose an angle threshold, e.g., thr = 2 degrees, and find strokes in ph
thr <- 2 / 180 * pi
zc_result <- zero_crossings(ph, thr, harbor_seal$M$sampling_rate / 0.2)
# positive-going half strokes
ps <- zc_result$K[zc_result$s>0] / harbor_seal$M$sampling_rate
# negative-going half strokes	
ns <- zc_result$K[zc_result$s<0] / harbor_seal$M$sampling_rate
}
