% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sm_xml_devid.R
\name{get_sm_xml_devid}
\alias{get_sm_xml_devid}
\title{Get SMRT tag device ID from xml file(s)}
\usage{
get_sm_xml_devid(xml_file)
}
\arguments{
\item{xml_file}{filename (with path) of xml file to parse, or xml document as already read in using xml2::read_xml}
}
\value{
a list with entries device_id (the ID as a hexadecimal string) and device_serial (the short ID -- an 8-digit number)
}
\description{
Parse xml file to get information about sensor configuration and tag metadata from SM board of SMRT tag
}
