% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssolve3.R
\name{lssolve3}
\alias{lssolve3}
\title{Least-squares solver for 3-dimensional data-driven calibration of field sensors.}
\usage{
lssolve3(X, G = diag(nrow = 3, ncol = 4), method = 1, Aux = NULL)
}
\arguments{
\item{X}{a 3-column data matrix representing measurements of a field vector (i.e., a constant norm). X may be affected by various calibration errors and by additive noise. The objective of this function is to infer the calibration errors in X so as to return an improved estimate of the correct field vector measurements.}

\item{G}{is the initial calibration matrix that was used to generate the incoming X. If X is uncalibrated, leave G at its default value.}

\item{method}{of calibration:  1 for offset only (the default), 2 for offset and gain, 3 for offset, gain, and cross-terms; or 4 for offset, gain, cross-terms, and an auxiliary covariate.}

\item{Aux}{A matrix of (optional) auxiliary co-variate(s) (e.g., temperature or pressure measurements) with the same number of rows as X. Each column in Aux is a covariate. If multiple covariates are provided, they should be reasonably uncorrelated to avoid numerical problems.}
}
\value{
A list with 2 elements:
\itemize{
\item{\strong{X: }} the improved data matrix after calibration errors have been corrected
\item{\strong{G: }}  a matrix of calibration corrections. The first three columns form a 3x3 matrix of gains and cross-terms. If method=1, this matrix will be the identity matrix. If method=2, the matrix will be diagonal. The next column of G is a vector of offsets. If an  auxiliary covariate is given, G will have an additional two columns for each column in Aux comprising the scale factor and offset of the covariate for each axis of X. If an input G is given, the output G will contain both the input and output calibrations as a single compound calibration.
}
}
\description{
This function is called by auto_cal_acc and auto_cal_mag. The function uses a locally linearized least-squares formulation so should be run iteratively several times until converged.
}
\examples{
lssolve3(beaked_whale$A$data)
}
