% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smrt.R
\name{read_smrt}
\alias{read_smrt}
\title{Read SMRT tag data file(s) and convert to .nc}
\usage{
read_smrt(
  data_dir = NULL,
  depid,
  sm_dir = paste(data_dir, "/SMfiles"),
  gps_file = NULL,
  archive_file = NULL,
  tz = "UTC",
  save_SAtimes = TRUE,
  SAtimes_fname = paste(depid, "_wav_filetimes.csv", sep = ""),
  nc_dir = getwd(),
  nc_fname = paste(depid, "_raw.nc", sep = ""),
  info = NULL,
  device_serial = NULL,
  device_url = "https://wildlifecomputers.com/"
)
}
\arguments{
\item{data_dir}{string containing the name (including full or relative path) of the directory where the data files are stored. If omitted, \code{fname} must include full or relative path information for the csv file(s).}

\item{depid}{string containing the deployment identification code assigned
to this deployment, for example, 'mn12_186a'. If \code{fname} is not input, csv files are assumed to have names of the form "\code{depid}_001.csv" (002, 003 etc if multiple files) or "\code{depid}.csv" (if data is in a single file).}

\item{sm_dir}{directory where files from the SM board of the SMRT tag are stored (these are the files "unpacked" from the .dtg files, and include .swv, .xml, .csv, and .wav files). If not provided, the defaul will be subdirectory "SMfiles" within \code{data_dir}.}

\item{gps_file}{name of file (including path) with tag GPS data. If not provided, this function will search for a file in \code{data_dir} which includes the string "FastGPS.csv".}

\item{archive_file}{name of file (including path) with tag "Archive" data from the Wildlife Computers board. If not provided, this function will search for a file in \code{data_dir} which includes the string "Archive.csv".}

\item{tz}{string indicating the time zone used by the SMRT device clock. Default: "UTC".}

\item{save_SAtimes}{Logical. Save a .csv file containing metadata about the acoustic recordings (.wav files) from \code{sm_dir}? Default: TRUE.}

\item{SAtimes_fname}{Filename to which to save the metadata about acoustic .wav file data. Defaults to: "[depid]_wav_filetimes.csv" stored in the current working directory.}

\item{nc_dir}{String containing the name (including full or relative path) of the directory where the output nc file should be stored. Defaults to the current working directory.}

\item{nc_fname}{String containing the file name to use for the output netCDF file. Defaults to "(depid)_raw.nc" - for example, "mn12_186a_raw.nc"}

\item{info}{a list of deployment-specific metadata (create using \link[tagtools]{make_info}, potentially with manual additions after creation). Default: make_info(tagtype = "SMRT", depid = depid). This becomes the basis for the info structure of the output NetCDF file.}

\item{device_serial}{String containing the serial number of the SMRT tag. Obtained from \code{txt_fname} or else defaults to NULL; stored in the info structure of the output NetCDF file.}

\item{device_url}{String containing URL of tag manufacturer; defaults to "https://wildlifecomputers.com/" and is stored in the info structure of the output NetCDF file.}
}
\value{
A string containing the file name of the netCDF (.nc) file in which the output has been saved. This function
generates a netCDF file in the current working directory containing
		the tag data variables, including:
		\itemize{
		\item {A, Accelerometer data structure}
		\item {M, Magnetometer data structure}
		\item {depth, Depth data structure}
		\item {temp, Temperature sensor data structure}
		\item {SA, Metadata about acoustic recordings (.wav files)}
		\item {info	Metadata about the deployment}
		}
}
\description{
Read data file(s) with data from a SMRT tag deployment, including associated metadata; apply initial calibration to convert to engineering units and store the resulting data in a .nc file.
}
\note{
SMRT tags can produce very large csv files which are slow to
process. This function is (somewhat) optimised for speed and memory use so will
tolerate large files. But processing could be slow. Note also that data from 3D sensors 
**will be in the NEU orientation** expected by the animaltag tool kit.
}
\examples{
\dontrun{
nc_filename <- read_smrt()
load_nc(nc_filename)
}
}
