% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smrt_archive.R
\name{read_smrt_archive}
\alias{read_smrt_archive}
\title{Read SMRT tag data from Wildlife Computers board (Archive file) and return list of sensor data lists}
\usage{
read_smrt_archive(depid, info, archive_file)
}
\arguments{
\item{depid}{string containing the deployment identification code assigned}

\item{info}{list with metadata about the deployment (used to obtain recording start time)}

\item{archive_file}{name of file (including path) with WC board data (required)}
}
\value{
A list of sensor data lists with sensor data, including:
		\itemize{
		\item {battery}
		\item {wet_dry, wet/dry sensor data (1 means dry)}
		}
}
\description{
Helper function for \link[tagtools]{read_smrt} - not ordinarily used independently. Read data data from Wildlife Computers board (Archive file) and return list of sensor data lists.
}
\examples{
\dontrun{
archive_data <- read_smrt_archive(depid, archive_file)
}
}
