% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smrt_gps.R
\name{read_smrt_gps}
\alias{read_smrt_gps}
\title{Read SMRT tag GPS data file and return list of sensor data lists}
\usage{
read_smrt_gps(depid, info, gps_file)
}
\arguments{
\item{depid}{Deployment ID string}

\item{info}{list with metadata about the deployment (used to obtain recording start time)}

\item{gps_file}{name of file (including path) with tag GPS data (required)}
}
\value{
A list of sensor data lists with GPS data, including:
		\itemize{
		\item {GPS_position, latitude and longitude}
		\item {GPS_satellites, number of satellites per position}
		\item {GPS_residual, GPS position residuals}
		\item {GPS_time_error, GPS time error}
		}
}
\description{
Helper function for \link[tagtools]{read_smrt} - not ordinarily used independently. Read data file with  GPS data from a SMRT tag deployment, including associated metadata.
}
\examples{
\dontrun{
gps_data <- read_smrt_gps(gps_file)
}
}
