% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nan.R
\name{remove_nan}
\alias{remove_nan}
\title{Remove missing values}
\usage{
remove_nan(x)
}
\arguments{
\item{x}{vector or matrix of numeric values}
}
\value{
x with NAs replaced as noted above
}
\description{
Replace any NaNs in the columns of X with the nearest non-NA number in the same column. If an entire column is NA, the first non-NA number in the matrix is used as a filler. If the entire matrix is NA, 1 is used as a filler.
}
\note{
Thanks to:  https://stackoverflow.com/questions/7735647/replacing-nas-with-latest-non-na-value
}
