% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_spectrum.R
\name{sens_spectrum}
\alias{sens_spectrum}
\title{Spectrum level of a sensor time series, i.e., the amount of power per 1 Hz band.}
\usage{
sens_spectrum(X, sampling_rate = NULL, nfft = 512, pct = 1, compute_ul = TRUE)
}
\arguments{
\item{X}{a sensor list or a vector or matrix containing the sensor signals to be processed. For multi-axial sensors when X is not a sensor list, each axis should be in a column of X.}

\item{sampling_rate}{is the sampling rate of the signals in X in Hz. This is only needed when X is not a sensor list.}

\item{nfft}{The length of the fft to use. Choose a power of two for fastest operation. Default value is 512.}

\item{pct}{optionally sets the percentage level used to compute the lower and upper spectra. The default value is 1 percent, which is suitable for 10s averages if there is at least 1 day of data. If the data duration is less than a day, pct should be increased accordingly if accurate spectra are required.}

\item{compute_ul}{Logical: whether to compute and output \code{Pu} and \code{Pl}, which give power in the intervals with highest and lowest activity. Default: TRUE}
}
\value{
A list including variables described below (u/l variables are only present if input \code{compute_ul} is TRUE.) P, Pu, and Pl will have the same number of columns as the input data X, if X is a matrix. Other elements are vectors whose length matches the number of frequencies, and the number of rows in P, Pu, Pl.
\itemize{
\item{\strong{P: }} is the spectral power at each frequency in power per Hz. The spectrum is single-sided and extends to \code{sampling_rate}/2. The spectrum is scaled so that the power in the signal is the same as the total power in the spectrum, i.e., \code{sampling_rate / nfft * sum(colSums(P))} should be equal to: \code{mean(colSums(X^2))}.
\item{\strong{freq: }} The vector of frequencies in Hz at which P is calculated.
\item{\strong{Pl: }} is the spectral power of nfft-length intervals with the lowest activity. Pl is scaled in the same way as P. 
\item{\strong{Pu: }} is the spectral power of nfft-length intervals with the highest activity. Pl is scaled in the same way as P. 
\item{\strong{tl: }} central times of the intervals listed in Pl
\item{\strong{tu: }} central times of the intervals listed in Pu
}
}
\description{
This replicates the function of Matlab's \code{psd} function (now deprecated in Matlab). 
The input signal is divided into overlapping pieces equal in length to the required Fast Fourier Transform (FFT) length.
Each piece is windowed and the FFT computed. 
The spectral power is then estimated from the mean of the spectral magnitudes squared. 
Power is scaled to account for the scale factor of the FFT and the window.
The power is also scaled by 10log10 of the bin width in Hz (i.e., the sampling rate divided by the FFT length) to convert the per-bin powers into approximate per-Hz powers.
}
\examples{
spect_data <- sens_spectrum(beaked_whale$M, nfft = 8)
}
