% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_assemble_swv.R
\name{sm_assemble_swv}
\alias{sm_assemble_swv}
\title{Assemble data from a series of swv files into sensor data lists}
\usage{
sm_assemble_swv(
  sm_dir,
  depid,
  df = 1,
  ch = NULL,
  recn = NULL,
  sm_file_info = sm_fnames(sm_dir, depid),
  xml_info = sm_get_config(sm_dir),
  sensor_defs = sm_channels(xml_info$unique_channels),
  discard = 1,
  quiet = TRUE
)
}
\arguments{
\item{sm_dir}{name of directory (including path) where SM data files (.swv and .csv files) are stored}

\item{depid}{Deployment ID string. Optional, but may be helpful to include if the directory where .swv files are stored contains data from more than one tag deployment.}

\item{df}{decimation factor. Default: 1 (no decimation). If a single df value is input, data will be decimated to give a sampling rate for each channel of 1/df of the full original sampling rate. df can also be a vector the same length as ch or the number of rows in \code{sensor_defs}, if different decimation factors are desired per sensor channel. Decimation is done via \code{\link{decz}} (which calls \code{\link{decdc}}), and includes application of a low-pass anti-alias filter and correction for the group delay of the filter (for "DC accuracy").}

\item{ch}{a vector of strings (e.g. 'acc', 'mag', 'pres') or channel numbers indicating which sensor channels to read data from. The channel numbers are the same as those used in the xml metadata files. Default: NULL (read all channels in the .swv file). If \code{sm_file_info} is input, the input \code{ch} should not include channels excluded from \code{sm_file_info}.}

\item{recn}{a numeric vector indicating which swv/csv files should be read. The record numbers are included in file names (last 3 digits), and can be obtained via \code{sm_fnames(sm_dir, depid)}. Default: all files present in sm_dir. This might be used to avoid reading in a long series of data recorded after a tag fell off, for example...but otherwise beware introducing synchronization errors between sensors -- probably best to read all data and then use \code{crop()} later...If \code{sm_file_info} is input, the input \code{recn} should not include files excluded from \code{sm_file_info}.}

\item{sm_file_info}{information about swv files in sm_dir. If not input, it is obtained via \code{sm_file_info <- sm_fnames(sm_dir, depid)}}

\item{xml_info}{metadata about sensor recordings from xml files in sm_dir. If not input, it is obtained via \code{xml_info <- sm_get_config(sm_dir)}.}

\item{sensor_defs}{metadata about sensors on the tag. If not input, it is obtained via \code{channel_info <- sm_channels(xml_info$unique_channels)}.}

\item{discard}{duration of data (in seconds) to discard (and replace with NA) at the start of recording and after each gap in data, to avoid power-up transients. Default: 1 second.}

\item{quiet}{logical; set to TRUE (the default) to suppress messages about "reading file..." You may want not-quiet operation to monitor progress if many large files are being read, slowly.}
}
\value{
A list containing one sensor data list for each sensor in \code{ch}, plus a list item named sampling_rate with the final sampling rate of each sensor
}
\description{
Data from the SM board of SMRT tags is stored in sensor wav files (.swv files). This function calls \code{\link{sm_parse_swv}} to read the data from individual files and uses timing information (obtained via a call to \code{\link{sm_cuetab}}) to assemble them into a continuous timeseries for each sensor, accounting for any recording gaps or timing errors. Data are returned in the form of one sensor data list per sensor processed. Different sensors may be sampled at different sampling rates.
}
\note{
This function is analogous in function to d3readswv from the DTAG Matlab tool kit. d3readswv() also has an option to return only sampling rate and sensor channel names instead of sensor data; for this function use \link{sm_get_config} and perhaps \link{sm_channels}.
}
