% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_ch_subset.R
\name{sm_ch_subset}
\alias{sm_ch_subset}
\title{Subset sensor_defs data.frame to include only sensor channels in ch}
\usage{
sm_ch_subset(sensor_defs, ch)
}
\arguments{
\item{sensor_defs}{metadata about sensors on the tag (likely obtained by a call to \code{\link{sm_channels}}).}

\item{ch}{a vector of strings (e.g. 'acc', 'mag', 'pres') or channel numbers indicating which sensor channels to read data from. The channel numbers are the same as those used in the xml metadata files. If no \code{ch} is input, this function will simply return the input \code{sensor_defs}.}
}
\value{
The \code{sensor_defs} data.frame, subsetted to include only the channels in \code{ch}.
}
\description{
This is a utility function called by other functions that work with SMRT data (not usually called directly by users)
}
