% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_cuetab.R
\name{sm_cuetab}
\alias{sm_cuetab}
\title{Get "cue tab" (information about file timing) for SM board wav or swv data files}
\usage{
sm_cuetab(
  sm_dir,
  sm_file_info = sm_fnames(sm_dir),
  xml_info = sm_get_config(sm_dir),
  err_thr_sec = 0.005,
  err_thr_samp = 10,
  suffix = "wav"
)
}
\arguments{
\item{sm_dir}{directory where data files from the SM board or DTAG (e.g., xml and swv files) are stored}

\item{sm_file_info}{a data frame with information about SM data file names in sm_dir. If not input, it is obtained by a call to \code{\link{sm_fnames}}; providing it as input may save a little time.}

\item{xml_info}{a list of metadata extracted from the SM xml files. If not input, it is obtained by a call to \code{\link{sm_get_config}}; providing it as input may save a little time.}

\item{err_thr_sec}{threshold in seconds for reporting timing errors (that also exceed err_thr_samp). Default: 0.005 sec.}

\item{err_thr_samp}{threshold in samples for reporting timing errors (that also exceed err_thr_sec). Default: 10 samples.}

\item{suffix}{file extension of the files to catalog. They must be in a wav-format configuration. Default: 'wav' (could also be 'swv' or some other file extension assigned in the future to files in wav format).}
}
\value{
a "cuetab" data.frame including variables:
		\itemize{
		\item {recn: recording number}
		\item {start_sec: start time of block, in seconds since start of recording}
		\item {n_samples: total number of samples in the block}
		\item {status: whether data in this block is data (status = 1), zero-filled (status = 0), or and unfilled gap (status = -1)}
		}
}
\description{
this function is called by other functions that read SM board data files. It is not normally called by end users.
}
