% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_fnames.R
\name{sm_fnames}
\alias{sm_fnames}
\title{Get SMRT SM data file names, given a directory to search and a deployment ID}
\usage{
sm_fnames(sm_dir, depid = "")
}
\arguments{
\item{sm_dir}{directory where xml file(s) are stored}

\item{depid}{deployment ID string}
}
\value{
A data.frame of information about data files in sm_dir including variables:
		\itemize{
		\item {device_serial: ID number of the tag}
		\item {file_name: base data file names (without file extensions like ".xml")}
		\item {recn: recording number}
		\item {sm_dir: file location (as input)}
		}
}
\description{
Get SMRT SM data file names, given a directory to search and a deployment ID
}
\note{
Unlike the Matlab dtag tools, this function will not store results in the temporary/working directory. Info is re-read from sm_dir each time the function is called.
}
