% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_get_cal.R
\name{sm_get_cal}
\alias{sm_get_cal}
\title{Get SMRT SM board sensor calibration info from xml file(s)}
\usage{
sm_get_cal(sm_dir = NULL, xml_file = NULL, attr_path = NULL)
}
\arguments{
\item{sm_dir}{directory where xml file(s) are stored}

\item{xml_file}{filename (with path) of xml file to parse. If provided, sm_dir will be ignored.}

\item{attr_path}{directory containing device-specific attribute files. Most SMRT devices do not have these, so if you don't know what it is or don't have any, leave \code{attr_path} at its default NULL value. In this case,  generic (factory preset values based on sensor specifications) will be used, which (in combination with data-driven calibrations) is sufficient in most cases.}
}
\value{
A "CAL" list of calibration constants and metadata including elements:
		\itemize{
		\item {PRESS}
		\item {TEMP}
		\item {ACC}
		\item {MAG}
		\item {AUDIO}
		}
}
\description{
Parse xml file(s) to get information about sensor calibration constants from SM board of SMRT tag
}
\note{
DTAG and Matlab compatibility: This function replicates some of the tasks of DTAG Matlab tools d4readattr and d4decodeattr.
}
