% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_parse_swv.R
\name{sm_parse_swv}
\alias{sm_parse_swv}
\title{Read in one swv file from a SMRT tag's SM board, and integrate information about file timing and gaps}
\usage{
sm_parse_swv(
  swv_file,
  ch = NULL,
  start_sec = NULL,
  end_sec = NULL,
  depid = NULL
)
}
\arguments{
\item{swv_file}{filename (with path) of swv file to be read}

\item{ch}{a vector of sensor numbers or names to read in. If ch is specified, only the sensor channels matching the type (or numbers) given will be read. To find out which channels are available in a dataset, use: \code{\link{sm_channels}}.}

\item{depid}{Deployment ID string. Optional, but may be helpful to include if the directory where .swv files are stored contains data from more than one tag deployment.}

\item{start_samp}{time (in samples since file start) to start reading data. Default: start of file.}

\item{end_samp}{time (in samples since file start) to start reading data. Default: end of file.}
}
\value{
A list containing metadata and sensor data vectors (for triaxial sensors, there will be one per axis). Each may be a different length according to the sampling rate of the sensor channel.
		\itemize{
		\item {data: sensor data vectors (each vector's length depends on sampling rate; number of entries in data matches length of sampling_rate and ch_nums)}
		\item {sampling_rate: vector of sampling rates. Length corresponds to the number of sensors represented in \code{data}.}
		\item {ch_nums: vector of channel ID numbers. Length corresponds to the number of sensors represented in \code{data}.}
		}
}
\description{
SM boards in SMRT tags store metadata about file timing, as well as storing data from sensors (sampled perhaps at different rates) in swv or "sensor wav" files. This function reads in one such .swv file. This function is not normally called by users, but instead called by read_smrt_sm(), which calls sm_read_swv(), which calls this function. (The file names and function call structure loosely mimic the DTAG Matlab tool kit of Mark Johnson, after which these tools are modeled.)
}
