% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_archive.R
\name{sound_archive}
\alias{sound_archive}
\title{Create a "sound archive" (SA) sensor data list with metadata about acoustic recordings from a SMRT or DTAG deployment.}
\usage{
sound_archive(
  sm_dir,
  depid,
  err_thr_sec = 0.005,
  err_thr_samp = 10,
  suffix = "wav"
)
}
\arguments{
\item{sm_dir}{directory where xml file(s) are stored}

\item{depid}{deployment ID string}

\item{err_thr_sec}{threshold in seconds for reporting timing errors (that also exceed err_thr_samp). Default: 0.005 sec.}

\item{err_thr_samp}{threshold in samples for reporting timing errors (that also exceed err_thr_sec). Default: 10 samples.}

\item{suffix}{file extension of the files to catalog. They must be in a wav-format configuration. Default: 'wav' (could also be 'swv' or some other file extension assigned in the future to files in wav format).}
}
\value{
A data.frame with metadata about acoustic wav files associated with the tag deployment. Variables include:
		\itemize{
		\item {recn: file number}
		\item {start_sec: time of first sample in block, in seconds since recording start}
		\item {n_samples: number of samples in the block}
		\item {status: status of the block. 0 means data; -1 means an unfilled gap; 1 means a zero-filled gap.}
		}
}
\description{
SMRT tags and DTAGs record acoustic data in wav files. These files are too large to include inside a netCDF file. However, we can include metadata about the recordings (file names, duration, sampling rate, and information about any gaps in recording) in a "SA" ("sound archive") list for inclusion in the netCDF archive. This function creates the SA list, which can also be used to figure out where (which file, and when within the file) to find a sound clip at a particular time.
}
