% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_ls.R
\name{spherical_ls}
\alias{spherical_ls}
\title{Deduce the calibration constants}
\usage{
spherical_ls(X, field_strength = NULL, cal, method = 1, Aux = NULL)
}
\arguments{
\item{X}{a 3-column data matrix representing measurements of a field vector (i.e., a constant norm). X may be affected by various calibration errors and by additive noise. The objective of this function is to infer the calibration errors in X so as to return an improved estimate of the correct field vector measurements.}

\item{field_strength}{is the target field strength in the same units as X.}

\item{cal}{is an optional list of calibration information. Only cal$poly and cal$cross are supported.}

\item{method}{of calibration:  1 for offset only (the default), 2 for offset and gain, 3 for offset, gain, and cross-terms; or 4 for offset, gain, cross-terms, and an auxiliary covariate.}

\item{Aux}{A matrix of (optional) auxiliary co-variate(s) (e.g., temperature or pressure measurements) with the same number of rows as X. Each column in Aux is a covariate. If multiple covariates are provided, they should be reasonably uncorrelated to avoid numerical problems.}
}
\value{
A list with elements:
\itemize{
\item{\strong{X: }} the improved data matrix after calibration errors have been corrected
\item{\strong{cal: }} the improved calibration list
\item{\strong{sigma: }}  a two-element vector reporting the standard deviation of field strength in the data, before and after the data-driven calibration
}
}
\description{
Least squares solver for spherical data-driven calibration. This function is used by \code{\link{auto_cal_acc}} and \code{\link{auto_cal_mag}}, and is a wrapper function for \code{\link{lssolve3}}.
}
\examples{
slso <- spherical_ls(beaked_whale$A$data, 9.81)
}
