% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagtools-package.R
\docType{package}
\name{tagtools-package}
\alias{tagtools-package}
\alias{tagtools}
\title{tagtools}
\description{
Work with Data from High-Resolution Biologging Tags
}
\seealso{
Useful links:
\itemize{
  \item \url{<https://animaltags.org>}
  \item \url{<https://animaltags.github.io/tagtools_r/index.html>}
  \item Report bugs at \url{https://github.com/animaltags/tagtools_r/issues}
}

}
\author{
\strong{Maintainer}: Stacy DeRuiter \email{stacy.deruiter@calvin.edu} (\href{https://orcid.org/0000-0002-0571-0306}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mark Johnson [copyright holder]
  \item David Sweeney
  \item Ye Joo McNamara-Oh
  \item Samuel Fynewever
  \item (Oghenkevwe) Racheal Tejevbo
  \item Tiago Marques
  \item Yuqian Wang
  \item (Oghenesuvwe) Su Ogedegbe
}

}
\keyword{internal}
