% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smrt_sm.R
\name{read_smrt_sm}
\alias{read_smrt_sm}
\title{Read SMRT tag data from SM board (.swv and .csv files) and return list of sensor data lists}
\usage{
read_smrt_sm(depid, info, sm_dir, ch = NULL)
}
\arguments{
\item{depid}{string containing the deployment identification code assigned}

\item{info}{list with metadata about the deployment (used to obtain recording start time)}

\item{sm_dir}{name of directory (including path) where SM data files (.swv and .csv files) are stored}

\item{ch}{a vector of strings (e.g. 'acc', 'mag', 'pres') or channel numbers indicating which sensor channels to read data from. The channel numbers are the same as those used in the xml metadata files. Default: NULL (read all channels in the .swv file).}
}
\value{
A list of sensor data lists with sensor data, including:
		\itemize{
		\item {A}
		}
}
\description{
Helper function for \link[tagtools]{read_smrt} - not ordinarily used independently. Read data from SM board (.swv files) and return list of sensor data lists.
}
\examples{
\dontrun{
sm_data <- read_smrt_sm(depid, info, sm_dir)
}
}
