% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_channels.R
\name{sm_channels}
\alias{sm_channels}
\title{Get metadata about SM board / DTAG3 / DTAG4 sensor channels}
\usage{
sm_channels(ch = NULL, sm_dir = NULL, depid = NULL)
}
\arguments{
\item{ch}{a vector of sensor numbers or names for which numbers, names, and/or descriptions are required. This input is optional; if not specified, information about all channels recorded by the tag is returned.}

\item{sm_dir}{directory where data files from the SM board or DTAG (e.g., xml and swv files) are stored}

\item{depid}{Deployment ID string (e.g., "Zica-20260927-12345" or "sw26_123a"). Optional - if not specified then all files in sm_dir will be included}
}
\value{
A data.frame of metadata including variables:
		\itemize{
		\item {ch_names: sensor channel names as strings}
		\item {ch_nums: numeric IDs for each sensor channel}
		\item {description: a description of each channel}
		}
}
\description{
SM boards in SMRT tags, and DTAG 3-4 tags, have a system of numeric and string identifiers for different sensors. This function allows conversion between numeric channel IDs and human-readable string IDs.
}
\note{
If no inputs are provided, sm_channels returns metadata about all sensor channels known. If ch is not input but sm_dir and depid are, then information about sensor channels available for deployment depid in tag data files in sm_dir will be read from the data files.
}
